% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.utils.R
\name{metadata_curl_runner}
\alias{metadata_curl_runner}
\title{Metadata Curl Runner}
\usage{
metadata_curl_runner(sessionID, URL, root, SOAPAction)
}
\arguments{
\item{sessionID}{a character string that is the sessionID element returned by \link{rforcecom.login}}

\item{URL}{a character string URL where the curlPerform will send the data}

\item{root}{XML document of metadata formatted to submit to the Metadata API}

\item{SOAPAction}{a charac indicating the SOAPAction to perform}
}
\value{
A XML document returned from running \code{curlPerform}
}
\description{
This function formats and runs requests to the Metadata API
}
\concept{
metadata salesforce api
}

