% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.utils.R
\name{metadataListToXML}
\alias{metadataListToXML}
\title{List to XML COnverter}
\usage{
metadataListToXML(root, sublist, metatype=NULL)
}
\arguments{
\item{root}{XML document serving as the basis upon which to add the list}

\item{sublist}{a \code{list} containing data to add to the root node}

\item{metatype}{a character indicating the element name of each record in the list}
}
\value{
A XML document with the sublist data added
}
\description{
This function converts a list of metadata to XML
}
\concept{
metadata salesforce api
}

