% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.utils.R
\name{query_parser}
\alias{query_parser}
\title{A SOQL Response XML Parser}
\usage{
query_parser(root)
}
\arguments{
\item{root}{XML document returned by a response from the curl request in \link{rforcecom.query}}
}
\value{
A \code{data.frame} containing the query result
}
\description{
This function parses the response from \link{rforcecom.query}
}
\concept{
query soql salesforce api
}

