% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.deployMetadata.R
\name{rforcecom.cancelDeployMetadata}
\alias{rforcecom.cancelDeployMetadata}
\title{Cancel a Prior Call to Deploy Metadata}
\usage{
rforcecom.cancelDeployMetadata(session, id, verbose=FALSE)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as 
returned by \link{rforcecom.login}}

\item{id}{a character string id returned from \link{rforcecom.deployMetadata}}

\item{verbose}{a boolean indicating whether to print the XML request}
}
\description{
This function cancels a deployment of metadata 
as a package XML files to a target environment.
}
\examples{
\dontrun{

deployOptions <- list(allowMissingFiles='true', performRetrieve='false')
deploy_info <- rforcecom.deployMetadata(session, 'package.zip', deployOptions)

# cancel deployment
deploy_cancel_info <- rforcecom.cancelDeployMetadata(session, deploy_info$id)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_canceldeploy.htm}
}
\concept{
cancel deploy metadata salesforce api
}

