% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.deployMetadata.R
\name{rforcecom.checkDeployStatusMetadata}
\alias{rforcecom.checkDeployStatusMetadata}
\title{Check on Deploy Calls and Get Contents If Available}
\usage{
rforcecom.checkDeployStatusMetadata(session, 
                                           id,
                                           includeDetails=c('false', 'true'),
                                           filename='deployed_package.zip', 
                                           verbose=FALSE)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as 
returned by \link{rforcecom.login}}

\item{id}{a character string id returned from \link{rforcecom.retrieveMetadata}}

\item{includeDetails}{a boolean. Set to false to check the status of the deploy without 
attempting to deploy details portion of the response. See the url for what is 
contained in the DeployDetails portion:
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_deployresult.htm#deploy_details_title}}

\item{filename}{a file path to save the zip file in the event that it is downloaded. The 
name must have a .zip extension. The default behavior will be to save in the current 
working directory as package.zip}

\item{verbose}{a boolean indicating whether to print messages during metadata creation}
}
\value{
A \code{list} of the response
}
\description{
This function returns details about an initiated deployMetadata requset
and saves the results into a zip file
}
\examples{
\dontrun{

deployOptions <- list(allowMissingFiles='true', performRetrieve='false')
deploy_info <- rforcecom.deployMetadata(session, 'package.zip', deployOptions)

# check on deploy status
deploy_status <- rforcecom.checkDeployStatusMetadata(session, deploy_info$id)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_checkdeploystatus.htm}
}
\concept{
deploy metadata salesforce api
}

