% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.retrieveMetadata.R
\name{rforcecom.checkRetrieveStatusMetadata}
\alias{rforcecom.checkRetrieveStatusMetadata}
\title{Check on Retrieve Calls and Get Contents If Available}
\usage{
rforcecom.checkRetrieveStatusMetadata(session, id, 
                                             includeZip=c('true', 'false'), 
                                             filename='package.zip', 
                                             verbose=FALSE)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as 
returned by \link{rforcecom.login}}

\item{id}{a character string id returned from \link{rforcecom.retrieveMetadata}}

\item{includeZip}{a boolean Set to false to check the status of the retrieval without 
attempting to retrieve the zip file. If omitted, this argument defaults to true, 
which means that the zip file is retrieved on the last call to checkRetrieveStatus() 
when the retrieval has finished.}

\item{filename}{a file path to save the zip file in the event that it is downloaded. The 
name must have a .zip extension. The default behavior will be to save in the current 
working directory as package.zip}

\item{verbose}{a boolean indicating whether to print messages during metadata creation}
}
\value{
A \code{list} of the response
}
\description{
This function returns details about an initiated retrieveMetadata requset
and saves the results into a zip file
}
\examples{
\dontrun{

retrieveRequest <- list(unpackaged=list(types=list(members='*', name='CustomObject')))
retrieve_info <- rforcecom.retrieveMetadata(session, retrieveRequest)

# check on status, this will automatically download the contents to package.zip if they are ready
retrieve_status <- rforcecom.checkRetrieveStatusMetadata(session, retrieve_info$id)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_checkretrievestatus.htm}
}
\concept{
retrieve metadata salesforce api
}

