% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.deployMetadata.R
\name{rforcecom.deployMetadata}
\alias{rforcecom.deployMetadata}
\title{Deploy Metadata to an Environment}
\usage{
rforcecom.deployMetadata(session, zipFile, deployOptions=NULL)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as 
returned by \link{rforcecom.login}}

\item{zipFile}{a file path to a .zip containing files for deployment}

\item{deployOptions}{a \code{list} of parameters defining the deployment}
}
\value{
A \code{list} of details from the created retrieve request
}
\description{
This function deploys metadata 
as a package XML files to a target environment.
}
\note{
See the Salesforce documentation for the accepted deployOptions parameters.
Here is a link to that documentation: 
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_deploy.htm#deploy_options}
}
\examples{
\dontrun{

deployOptions <- list(allowMissingFiles='true', performRetrieve='false')

deploy_info <- rforcecom.deployMetadata(session, deployOptions, 'package.zip')

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_deploy.htm}
}
\concept{
deploy metadata salesforce api
}

