% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.deployMetadata.R
\name{rforcecom.deployRecentValidation}
\alias{rforcecom.deployRecentValidation}
\title{ReDeploy a Prior Call to Deploy Metadata}
\usage{
rforcecom.deployRecentValidation(session, id)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as 
returned by \link{rforcecom.login}}

\item{validationID}{a character string id returned from \link{rforcecom.deployMetadata}}
}
\description{
This function redeploys a deployment of metadata 
to production in less time by skipping the execution of Apex tests.
}
\note{
Before you call this function your organization must have a validation that was 
recently run. You can run a validation on a set of components by calling \link{rforcecom.deployMetadata} 
with the checkOnly property of the deployOptions parameter set to true.
}
\examples{
\dontrun{

# Before you call rforcecom.deployRecentValidation(), your organization must have a validation that was 
# recently run. You can run a validation on a set of components by calling rforcecom.deployMetadata() 
# with the checkOnly property of the deployOptions parameter set to true. 
deployOptions <- list(allowMissingFiles='true', checkOnly='true', performRetrieve='false')
deploy_info <- rforcecom.deployMetadata(session, 'package.zip', deployOptions)

# redeploy deployment
redeploy_info <- rforcecom.deployRecentValidation(session, deploy_info$id)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_deployRecentValidation.htm}
}
\concept{
deploy recent metadata salesforce api
}

