% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.describeMetadata.R
\name{rforcecom.describeMetadata}
\alias{rforcecom.describeMetadata}
\title{Describe the Metadata in an Organization}
\usage{
rforcecom.describeMetadata(session)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as returned by \link{rforcecom.login}}
}
\value{
A \code{data.frame}
}
\description{
This function returns details about the organization metadata
}
\examples{
\dontrun{

# describe metadata for the organization associated with the session
metadata_info <- rforcecom.describeMetadata(session)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
\concept{
describe metadata salesforce api
}

