\name{rforcecom.getObjectList}
\alias{rforcecom.getObjectList}

\title{
Retrieve a list of objects
}

\description{
Retrieve a list of objects
}

\usage{
rforcecom.getObjectList(session)
}

\arguments{
  \item{session}{ Session data. It can be retrieve from \code{\link{rforcecom.login}}. }
}

\value{
A list of objects
}

\author{
Takekatsu Hiramura <thira@plavox.info>
}

\examples{
\dontrun{
 # Retrieve a list of objects
 rforcecom.getObjectList(session)
 }
}

\keyword{connection}


