% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.listMetadata.R
\name{rforcecom.listMetadata}
\alias{rforcecom.listMetadata}
\title{List All Objects of a Certain Metadata Type in Salesforce}
\usage{
rforcecom.listMetadata(session, 
                              queries, 
                              asOfVersion=NULL, 
                              verbose=FALSE)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as returned by \link{rforcecom.login}}

\item{queries}{a \code{list} of \code{list}s with each element consisting of 2 components: 1) 
the metadata type being requested and 2) the folder associated with the type that required for types 
that use folders, such as Dashboard, Document, EmailTemplate, or Report.}

\item{asOfVersion}{a numeric specifying the API version for the metadata 
listing request. If you don't specify a value in this field, it defaults to the API 
version specified when you logged in. This field allows you to override the default 
and set another API version so that, for example, you could list the metadata for a 
metadata type that was added in a later version than the API version specified when you logged in.}

\item{verbose}{a boolean indicating whether to print the XML request}
}
\value{
A \code{data.frame} containing the queried metadata types
}
\description{
This function takes a query of metadata types and returns a 
summary of all objects in salesforce of the requested types
}
\note{
Only 3 queries can be specifed at one time, so the list length must not exceed 3.
}
\examples{
\dontrun{

# pull back a list of all Custom Objects and Email Templates
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
\concept{
list metadata salesforce api
}

