% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.readMetadata.R
\name{rforcecom.readMetadata}
\alias{rforcecom.readMetadata}
\title{Read Object or Field Metadata from Salesforce}
\usage{
rforcecom.readMetadata(session, 
                              metadata_type=c('CustomObject', 'CustomField'), 
                              object_names, 
                              verbose=FALSE)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as returned by \link{rforcecom.login}}

\item{metadata_type}{a character string on what type of metadata that is being read}

\item{object_names}{a character vector of names that we wish to read metadata for}

\item{verbose}{a boolean indicating whether to print messages during metadata creation}
}
\value{
A \code{list} containing a response for each requested object
}
\description{
This function takes a a request of named elements in Salesforce and 
returns their metadata
}
\examples{
\dontrun{

metadata_info <- rforcecom.readMetadata(session, 
                                        metadata_type='CustomObject', 
                                        object_names=c('Account'))
}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
\concept{
read metadata salesforce api
}

