% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.renameMetadata.R
\name{rforcecom.renameMetadata}
\alias{rforcecom.renameMetadata}
\title{Rename Metadata Elements in Salesforce}
\usage{
rforcecom.renameMetadata(session, 
                                metadata_type, 
                                oldFullname, newFullname, verbose=FALSE)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as returned by \link{rforcecom.login}}

\item{metadata_type}{a character string on what type of metadata to create}

\item{oldFullname}{a character string corresponding to the fullName of the element you would 
like to rename}

\item{newFullname}{a character string corresponding to the new fullName you would like 
to apply the targeted element}

\item{verbose}{a boolean indicating whether to print the XML request}
}
\value{
A \code{data.frame} containing the creation result for each submitted metadata component
}
\description{
This function takes an old and new name for a 
metadata element in Salesforce and applies the new name
}
\examples{
\dontrun{

renamed_custom_object <- rforcecom.renameMetadata(session, 
                                                  metadata_type='CustomObject', 
                                                  oldFullname='Custom_Account23__c', 
                                                  newFullname='Custom_Account24__c', verbose=FALSE)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
\concept{
rename metadata salesforce api
}

