% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforcecom.retrieveMetadata.R
\name{rforcecom.retrieveMetadata}
\alias{rforcecom.retrieveMetadata}
\title{Make A Request to Retrieve the Metadata}
\usage{
rforcecom.retrieveMetadata(session)
}
\arguments{
\item{session}{a named character vector defining parameters of the api connection as 
returned by \link{rforcecom.login}}

\item{retrieveRequest}{a \code{list} of parameters defining what XML file representations
should be returned}
}
\value{
A \code{list} of details from the created retrieve request
}
\description{
This function makes a request to retrieve metadata 
as a package XML files that can be modified and later
deployed into an environment
}
\note{
See the Salesforce documentation for the proper arguments to create a 
retrieveRequest. Here is a link to that documentation: 
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_retrieve_request.htm}
}
\examples{
\dontrun{

retrieveRequest <- list(unpackaged=list(types=list(members='*', name='CustomObject')))

metadata_info <- rforcecom.retrieveMetadata(session, retrieveRequest)

}
}
\references{
\url{https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/}
}
\concept{
retrieve metadata salesforce api
}

